fix(test): move #1389 measurement hooks to root e2e conftest
The prior fold placed the ``pytest_sessionfinish`` / ``pytest_testnodedown``
/ ``pytest_terminal_summary`` hooks at
``tests/src/e2e/workflows/automation/conftest.py``. CI re-ran cleanly,
test passed, but the measurement section still didn't surface — the
xdist master only loads conftests up to where tests are collected, and
``pytest_terminal_summary`` (which writes to the visible terminal) is a
master-only hook. Deep subdir conftests don't get their hooks invoked
on the master.

The Readiness gate timings pattern works because its hooks live at
``tests/src/e2e/conftest.py`` (the root e2e conftest the master always
loads). Same fix here: move the storage, the
``record_poll_cadence_measurement`` helper, and the three hook
extensions to the root conftest alongside the readiness ones. The deep
``workflows/automation/conftest.py`` is removed; the measurement test
imports the helper from the root.
